Další smrsknutí importovaného: ll, FOR místo for (int i=0; i<N; i++)

+++ src/import/numth/composite_comb.cpp
@@ -1,13 +1,13 @@
 #define MAXC 111111
 #define MAX_PRIMES 20
-long long MOD = 1000000007;
+ll MOD = 1000000007;
 int fac[MAX_PRIMES][MAXC], ifac[MAX_PRIMES][MAXC];
 
 int PRIMES = 4, primes[5] = {3, 11, 29, 1609261};
 
-long long get_comb(int n, int k, int p) {
+ll get_comb(int n, int k, int p) {
     if (n < k) return 0;
-    long long ans = 1;
+    ll ans = 1;
     int pr = primes[p];
     while (n) {
         // Lucas theorem (binomial mod prime) - decompose n, k on digits in base p
@@ -19,10 +19,10 @@ long long get_comb(int n, int k, int p) {
     return ans;
 }
 
-long long get_comb(int n, int k) {
-    long long ans = 0, mod = 1;
+ll get_comb(int n, int k) {
+    ll ans = 0, mod = 1;
     FOR(p, PRIMES) {
-        long long a1 = ans, n1 = mod, a2 = get_comb(n, k, p), n2 = primes[p];
+        ll a1 = ans, n1 = mod, a2 = get_comb(n, k, p), n2 = primes[p];
         mod = n1 * n2;
         // Chinese remainder theorem
         // x = a1 (mod n1) && x = a2 (mod n2)  =>  x = ans (mod n1 * n2)
@@ -40,9 +40,9 @@ int composite_comb(int N, int K, int mod) {
     FOR(p, PRIMES) { // Precalculates factorials and inverses
         int pr = primes[p];
         fac[p][0] = 1;
-        FOR(i, pr) fac[p][i + 1] = ((long long)fac[p][i] * (i+1)) % pr;
+        FOR(i, pr) fac[p][i + 1] = ((ll)fac[p][i] * (i+1)) % pr;
         ifac[p][pr - 1] = power(fac[p][pr - 1], pr - 2, pr);
-        per(i, pr - 1, 0) ifac[p][i] = ((long long)ifac[p][i+1]*(i+1)) % pr;
+        per(i, pr - 1, 0) ifac[p][i] = ((ll)ifac[p][i+1]*(i+1)) % pr;
     }
 
     int ans = 0;
@@ -50,5 +50,5 @@ int composite_comb(int N, int K, int mod) {
         int n = min(K, i), k = N - 1;
         ans = (ans + get_comb(n, k)) % MOD;
     }
-    return (((long long)ans * N) % MOD * N) % MOD;
+    return (((ll)ans * N) % MOD * N) % MOD;
 }
